🎨 Refactor: 선호도 선택페이지 리팩토링
