+++ daeng/src/components/preference/PreferenceEdit.jsx
@@ -1,91 +1,39 @@
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { placeFeatures, placeTypes } from "../../data/CommonCode";
-import styled from "styled-components";
-import ConfirmBtn from "../commons/ConfirmBtn";
-import PreferencePlaceOption from "../commons/PreferencePlaceOption";
-import PreferenceFavoriteOption from "../commons/PreferenceFavoriteOption";
-import AlertDialog from "../commons/SweetAlert";
-import Loading from "../commons/Loading"; // Loading 컴포넌트 추가
-import restaurantIcon from "../../assets/icons/restaurant.svg";
-import cafeIcon from "../../assets/icons/cafe.svg";
-import parkIcon from "../../assets/icons/park.svg";
-import houseIcon from "../../assets/icons/house.svg";
-import galleryIcon from "../../assets/icons/gallery.svg";
-import playgroundIcon from "../../assets/icons/playground.svg";
-import travelIcon from "../../assets/icons/travel.svg";
-import museumIcon from "../../assets/icons/museum.svg";
-import filmIcon from "../../assets/icons/film.svg";
-import dogfood from "../../assets/icons/dogFood.svg";
-import run from "../../assets/icons/run.svg";
-import water from "../../assets/icons/water.svg";
-import toilet from "../../assets/icons/toilet.svg";
-import bug from "../../assets/icons/bug.svg";
-import cage from "../../assets/icons/cage.svg";
-import dogFriend from "../../assets/icons/dogFriend.svg";
-import paperbag from "../../assets/icons/paperbag.svg";
-import clean from "../../assets/icons/clean.svg";
-import gongwon from "../../assets/icons/gongwon.svg";
-import parkingLot from "../../assets/icons/parkingLot.svg";
 import axios from "axios";
+import AlertDialog from "../commons/SweetAlert";
+import PreferenceForm from "./PreferenceForm";
+import { placeTypes, placeFeatures } from "../../data/CommonCode";
+import { placeIcons, featureIcons } from "../../data/PreferenceIcons";
 
 function PreferenceEdit() {
   const navigate = useNavigate();
-  const [isLoading, setIsLoading] = useState(false); // 로딩 상태 추가
+  const [isLoading, setIsLoading] = useState(false);
   const [selectedPlaceOptions, setSelectedPlaceOptions] = useState([]);
   const [selectedFavoriteOptions, setSelectedFavoriteOptions] = useState([]);
 
-  const featureIcons = {
-    PLACE_FTE_01: dogfood,
-    PLACE_FTE_02: run,
-    PLACE_FTE_03: water,
-    PLACE_FTE_04: toilet,
-    PLACE_FTE_05: bug,
-    PLACE_FTE_06: cage,
-    PLACE_FTE_07: dogFriend,
-    PLACE_FTE_08: paperbag,
-    PLACE_FTE_09: clean,
-    PLACE_FTE_10: gongwon,
-    PLACE_FTE_11: parkingLot,
-  };
-
-  const placeIcons = {
-    PLACE_TYP_01: restaurantIcon,
-    PLACE_TYP_02: cafeIcon,
-    PLACE_TYP_03: parkIcon,
-    PLACE_TYP_04: houseIcon,
-    PLACE_TYP_05: playgroundIcon,
-    PLACE_TYP_06: travelIcon,
-    PLACE_TYP_07: galleryIcon,
-    PLACE_TYP_08: museumIcon,
-    PLACE_TYP_09: filmIcon,
-  };
-
   useEffect(() => {
     const fetchPreferences = async () => {
-      setIsLoading(true); // 로딩 시작
+      setIsLoading(true);
       try {
         const response = await axios.get(
           "https://dev.daengdaeng-where.link/api/v1/preferences",
           { withCredentials: true }
         );
-
         const data = response.data.data || [];
 
         const placeTypesMapping =
           data
             .find((item) => item.preferenceInfo === "시설종류")
             ?.preferenceTypes.map(
-              (label) =>
-                placeTypes.find((option) => option.name === label)?.codeId
+              (label) => placeTypes.find((option) => option.name === label)?.codeId
             ) || [];
 
         const placeFeaturesMapping =
           data
             .find((item) => item.preferenceInfo === "시설특징")
             ?.preferenceTypes.map(
-              (label) =>
-                placeFeatures.find((option) => option.name === label)?.codeId
+              (label) => placeFeatures.find((option) => option.name === label)?.codeId
             ) || [];
 
         setSelectedPlaceOptions(placeTypesMapping);
@@ -99,28 +47,11 @@ function PreferenceEdit() {
           confirmText: "확인",
         });
       } finally {
-        setIsLoading(false); // 로딩 종료
+        setIsLoading(false);
       }
     };
-
     fetchPreferences();
   }, []);
-
-  const handleOptionClick = (selectedOptions, setSelectedOptions, code) => {
-    if (selectedOptions.length >= 3 && !selectedOptions.includes(code)) {
-      AlertDialog({
-        mode: "alert",
-        title: "선택 초과",
-        text: "최대 3개만 선택 가능합니다.",
-        confirmText: "확인",
-      });
-      return;
-    }
-
-    setSelectedOptions((prev) =>
-      prev.includes(code) ? prev.filter((option) => option !== code) : [...prev, code]
-    );
-  };
 
   const handleConfirm = async () => {
     if (selectedPlaceOptions.length === 0 || selectedFavoriteOptions.length === 0) {
@@ -143,7 +74,7 @@ function PreferenceEdit() {
       preferenceTypes: selectedFavoriteOptions,
     };
 
-    setIsLoading(true); // 로딩 시작
+    setIsLoading(true);
     try {
       await axios.put(
         "https://dev.daengdaeng-where.link/api/v1/preferences",
@@ -180,135 +111,23 @@ function PreferenceEdit() {
         confirmText: "확인",
       });
     } finally {
-      setIsLoading(false); // 로딩 종료
+      setIsLoading(false);
     }
   };
-  
+
   return (
-    <Wrap>
-      {isLoading && <Loading label="로딩 중입니다..." />}
-      <Section>
-        <Title>어떤 시설에 관심이 많으신가요?</Title>
-        <StyledParagraph>* 최소 1개 ~ 3개 선택가능</StyledParagraph>
-        <OptionContainer>
-          {placeTypes.map(({ codeId, name }) => (
-            <PreferencePlaceOption
-              key={codeId}
-              label={name}
-              icon={placeIcons[codeId]} 
-              isSelected={selectedPlaceOptions.includes(codeId)}
-              onClick={() => handleOptionClick(selectedPlaceOptions, setSelectedPlaceOptions, codeId)}
-            />
-          ))}
-        </OptionContainer>
-      </Section>
-
-      <Section>
-        <Title>어떤 부분이 중요하신가요?</Title>
-        <StyledParagraph>* 최소 1개 ~ 3개 선택가능</StyledParagraph>
-        <OptionContainer>
-          {placeFeatures.map(({ codeId, name }) => (
-            <PreferenceFavoriteOption
-              key={codeId}
-              label={name}
-              icon={featureIcons[codeId]}
-              isSelected={selectedFavoriteOptions.includes(codeId)}
-              onClick={() =>
-                handleOptionClick(selectedFavoriteOptions, setSelectedFavoriteOptions, codeId)
-              }
-            />
-          ))}
-        </OptionContainer>
-      </Section>
-
-      <StyledParagraph2>
-        보호자님과 우리 댕댕이 맞춤 장소 추천을 위해 필요한 정보입니다.
-      </StyledParagraph2>
-
-      <Footer>
-        <ConfirmBtn label="수정 완료" onClick={handleConfirm} />
-      </Footer>
-    </Wrap>
+    <PreferenceForm
+      selectedPlaceOptions={selectedPlaceOptions}
+      setSelectedPlaceOptions={setSelectedPlaceOptions}
+      selectedFavoriteOptions={selectedFavoriteOptions}
+      setSelectedFavoriteOptions={setSelectedFavoriteOptions}
+      isLoading={isLoading}
+      onConfirm={handleConfirm}
+      confirmButtonLabel="수정 완료"
+      placeIcons={placeIcons}
+      featureIcons={featureIcons}
+    />
   );
 }
-
-const Wrap = styled.div`
-  padding: 20px;
-
-  @media (max-width: 554px) {
-    padding-left: 20px;
-    padding-right: 20px;
-    padding-top: 0px;
-  }
-`;
-
-const Section = styled.div`
-  margin-bottom: 40px;
-`;
-
-const Title = styled.h3`
-  text-align: left;
-  margin-left: 10px;
-  font-size: 17px;
-  margin-bottom: 10px;
-
-  @media (max-width: 554px) {
-    font-size: 14px;
-    margin-left: 20px;
-  }
-`;
-
-const StyledParagraph = styled.p`
-  font-size: 12px;
-  color: #ff69a9;
-  font-weight: bold;
-  margin-top: 9px;
-  margin-left: 10px;
-  display: flex;
-
-  @media (max-width: 554px) {
-    font-size: 10px;
-    margin-left: 20px;
-  }
-`;
-
-const OptionContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  margin-top: 10px;
-  gap: 5px;
-  margin-left: 5px;
-
-  @media (max-width: 554px) {
-    gap: 3px;
-  }
-`;
-
-const StyledParagraph2 = styled.p`
-  font-size: 14px;
-  color: red;
-  font-weight: bold;
-  margin-top: 80px;
-  margin-bottom: 20px;
-  display: flex;
-  justify-content: center;
-
-  @media (max-width: 554px) {
-    font-size: 10px;
-    margin-left: 10px;
-    text-align: center;
-    margin-top: 60px;
-  }
-`;
-
-const Footer = styled.div`
-  margin-left: 15px;
-  padding-bottom: 10px;
-
-  @media (max-width: 554px) {
-    margin-left: 20px;
-  }
-`;
 
 export default PreferenceEdit;
