+++ daeng/src/data/PreferenceIcons.js
@@ -1,0 +1,46 @@
+import restaurantIcon from "../assets/icons/restaurant.svg";
+import cafeIcon from "../assets/icons/cafe.svg";
+import parkIcon from "../assets/icons/park.svg";
+import houseIcon from "../assets/icons/house.svg";
+import galleryIcon from "../assets/icons/gallery.svg";
+import playgroundIcon from "../assets/icons/playground.svg";
+import travelIcon from "../assets/icons/travel.svg";
+import museumIcon from "../assets/icons/museum.svg";
+import filmIcon from "../assets/icons/film.svg";
+import dogfood from "../assets/icons/dogFood.svg";
+import run from "../assets/icons/run.svg";
+import water from "../assets/icons/water.svg";
+import toilet from "../assets/icons/toilet.svg";
+import bug from "../assets/icons/bug.svg";
+import cage from "../assets/icons/cage.svg";
+import dogFriend from "../assets/icons/dogFriend.svg";
+import paperbag from "../assets/icons/paperbag.svg";
+import clean from "../assets/icons/clean.svg";
+import gongwon from "../assets/icons/gongwon.svg";
+import parkingLot from "../assets/icons/parkingLot.svg";
+
+export const featureIcons = {
+  PLACE_FTE_01: dogfood,
+  PLACE_FTE_02: run,
+  PLACE_FTE_03: water,
+  PLACE_FTE_04: toilet,
+  PLACE_FTE_05: bug,
+  PLACE_FTE_06: cage,
+  PLACE_FTE_07: dogFriend,
+  PLACE_FTE_08: paperbag,
+  PLACE_FTE_09: clean,
+  PLACE_FTE_10: gongwon,
+  PLACE_FTE_11: parkingLot,
+};
+
+export const placeIcons = {
+  PLACE_TYP_01: restaurantIcon,
+  PLACE_TYP_02: cafeIcon,
+  PLACE_TYP_03: parkIcon,
+  PLACE_TYP_04: houseIcon,
+  PLACE_TYP_05: playgroundIcon,
+  PLACE_TYP_06: travelIcon,
+  PLACE_TYP_07: galleryIcon,
+  PLACE_TYP_08: museumIcon,
+  PLACE_TYP_09: filmIcon,
+};
